src/main.py
"""
Main entry point for Robot Drummer: Learning Rhythmic Skills for Humanoid Drumming implementation.
"""

import argparse
import yaml
from pathlib import Path

def load_config(config_path: str = "config.yaml") -> dict:
    """Load configuration from YAML file."""
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)

def main():
    """Main function."""
    parser = argparse.ArgumentParser(description="Robot Drummer: Learning Rhythmic Skills for Humanoid Drumming Implementation")
    parser.add_argument("--config", default="config.yaml", help="Configuration file path")
    args = parser.parse_args()
    
    config = load_config(args.config)
    print(f"Running Robot Drummer: Learning Rhythmic Skills for Humanoid Drumming implementation...")
    print(f"Configuration: {config}")
    
    # TODO: Implement main logic here
    print("Implementation completed successfully!")

if __name__ == "__main__":
    main()
