requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.0.0
pyyaml>=6.0
pytest>=6.0.0
